requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
anyio==4.11.0
APScheduler==3.10.4
attrs==25.4.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==43.0.1
frozenlist==1.8.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.1.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
multidict==6.7.0
oauth2client==4.1.3
oauthlib==3.3.1
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.2.5
python-dotenv==1.1.1
python-telegram-bot==21.4
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
yarl==1.22.0
